requirements.txt
torch
torchvision
pyyaml
mlflow